requirements.txt
pillow
scipy
websocket-client
pynput
wmi
requests